Extract SingularitySelectCard to its own file, byte-for-byte (#2326)
The pure move, so the repaint that follows has nothing structural in it.
`SingularitySelectCard` leaves `FactionSelectCard.tsx` for
`components/selectCard/SingularitySelectCard.tsx`, matching how
`components/taskCard/` is laid out and how #2322 laid out S.N.I.D.E.'s.
The manifest row keeps its shape — `lazyArchetype(() => import(...))` now
points at the new module and drops the
`.then((m) => ({ default: m.SingularitySelectCard }))` unwrap, because the
new file default-exports it the way every task card does. Registration
lives in `factions/singularity.ts`, so no sibling child touches it.

THE `TERM_*` CONSTS MOVE WITH THE CARD, which is the hazard #2326 names:
they are declared ABOVE the function, so a line-range cut leaves them in
the dispatcher or hands them to the Everymen tile that follows. All five —
`TERM_VOID` / `TERM_GREEN` / `TERM_BLUE` / `TERM_CYAN` / `TERM_CTA_BG` —
are read by this tile and nothing else, verified by grep across
`frontend/src`, so all five move.

`CAST` DOES NOT MOVE. It is shared — the Everymen tile (#2327) reads it
too — and its declaration carries the dispatcher's "the tiles are NOT dark"
record, which is a property of the grid rather than of any one tile. Rather
than export it (a shared line every sibling would then rebase against), the
one site here inlines the string it already resolved to:
`0 8px 26px var(--color-cast-shadow)`. Same characters, same markup; the
next commit retires the site anyway for the register's own lift token.

PROOF THE DIRECTORY IS UNCHANGED. `renderToStaticMarkup` of
`<FactionSelectCard faction="singularity" />` across all six state x members
combinations is byte-identical before and after (22,098 bytes, `diff`
clean). The probe was a throwaway; the standing guard is
`factionSelectCardDispatch.test.tsx`, green.

ONE PARAGRAPH OF THE DOCSTRING IS DELETED RATHER THAN CARRIED. It described
`const green = "#4ade80"` reaching "four style objects and two hover
handlers as an Identifier" in the present tense. #1609 retired that
constant; the prose was the last trace of it in the repo, and a decision
record about a file's neighbours ("the S.N.I.D.E. tile that then stood
directly above it") stops being true the moment the file is one tile long.
The live half — that these are `var()` strings, and why the family is
theme-invariant — stays. The dispatcher's "the tiles are dark" was already
deleted by #2322 and is not restored.
